pinger_server/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
mysqlclient==1.4.6
python-dotenv==0.13.0
djoser==2.0.3